+++ src/Rask.Wasm/Browser/RaskWasmBrowserJsonContext.cs
@@ -15,4 +15,5 @@ namespace Rask.Wasm.Browser;
 [JsonSerializable(typeof(PushSubscription))]
 [JsonSerializable(typeof(NotificationOptions))]
 [JsonSerializable(typeof(OrientationReading))]
+[JsonSerializable(typeof(IdleReading))]
 internal sealed partial class RaskWasmBrowserJsonContext : JsonSerializerContext;